fix(theme): add fontWeightStrong to allowedAntdTokens for bold markdown rendering
**Problem**: Bold markdown (`**text**`) wasn't rendering with proper font-weight in dashboard widgets after Ant Design v5 upgrade (issue #35485)

**Root Cause**:
- `fontWeightStrong` token was missing from `allowedAntdTokens` array
- Theme.getFilteredAntdTheme() filtered it out
- GlobalStyles.tsx:66 set `font-weight: undefined` on `<strong>` tags
- Browser fell back to user-agent stylesheet instead of theme value

**Solution**: Add `'fontWeightStrong'` to allowedAntdTokens in types.ts

**Testing**:
- Added test verifying fontWeightStrong exists in theme
- Test failed before fix (undefined), passes after fix
- All 15 Markdown tests pass
- Converted Markdown.test.jsx to TypeScript with proper types

**Impact**: Fixes bold rendering for all `<strong>` tags including markdown, restoring theme-controlled font weights
