+++ superset-frontend/packages/superset-ui-core/src/theme/types.ts
@@ -268,6 +268,7 @@ export const allowedAntdTokens = [
   'controlTmpOutline',
   'fontFamily',
   'fontFamilyCode',
+  'fontWeightStrong',
   'fontHeight',
   'fontHeightLG',
   'fontHeightSM',
